monotonic clock: Wrap a clock_difftime with clock_difftime_safe
The _safe version handles wraparounds for those platforms where a monotonic
clock is not available. This doesn't completely solve the issue but it handles
the most likely cases.

It doesn't the case where the clock is always set to the same time for example
but it does handle the case where the clock is corrected from 2020 to 2012.